pyproject.toml
[tool.poetry]
name = "bird-api"
version = "0.1.0"
description = "Api for a efficiant-net trained on the 525 birds dataset"
authors = ["Guillaume Ruff <guiguiruff@gmail.com>", "Benjamin Crouts De Paille <benjamin.crouts-de-paille@utbm.fr>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
tensorflow = "2.13.0"
Flask = "^3.0.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
